requirements.txt
ccxt
kafka-python
pandas
numpy
statsmodels
pyarrow
fastparquet
python-dotenv